+++ hume-ops/src/text_object/tests/mod.rs
@@ -1,5 +1,6 @@
 mod argument;
 mod bracket;
 mod line;
+mod paragraph;
 mod quote;
 mod word;